requirements.txt
asgiref==3.5.2
build==0.8.0
click==8.1.3
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==5.0.4
fonttools==4.37.3
fpdf2==2.5.7
joblib==1.2.0
mccabe==0.7.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pep517==0.13.0
Pillow==9.2.0
pip-tools==6.8.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
svg.path==6.2
threadpoolctl==3.1.0
tomli==2.0.1
whitenoise==6.2.0
